Add release script to generate zip artefacts
This is the recommended deployment method for Azure Functions, so let's
start supporting that instead of using the bloated and annoying `func`
CLI, which is a trial to get working every time I need to use it.

This also gives us an artefact we can publish on a Github release tag.